wip: moved around types + styling
