+++ extensions/vscode/src/integrations/creator/index.ts
@@ -1,107 +1,10 @@
 import * as vscode from 'vscode';
 import { assert } from '../../util/assert';
-import { ChatMessage } from 'core';
-import { Core } from 'core/core';
+import { CreatorModeState, ExecutePlanRequest, IPearAICreatorMode } from 'core';
 import { IMessenger } from 'core/util/messenger';
 import { ToCoreFromIdeOrWebviewProtocol } from 'core/protocol/core';
 import { FromCoreProtocol } from 'core/protocol';
 
-type CreatorModeState = "OVERLAY_CLOSED" | "OVERLAY_OPEN" | "OVERLAY_CLOSED_CREATOR_ACTIVE";
-
-/**
- * Interface for the Creator Mode API
- * Provides methods and events for controlling the Creator Mode UI and functionality
- */
-export interface IPearAICreatorMode {
-  /**
-   * Event that fires when the creator mode is activated or deactivated
-   */
-  readonly onDidChangeCreatorModeState: vscode.Event<CreatorModeState>;
-
-  /**
-   * Event that fires when a new task should be created
-   * Other extensions can listen to this event to handle task creation
-   */
-  readonly onDidRequestNewTask: vscode.Event<CreatorTaskRequest>;
-
-  /**
-   * Event that fires when a plan has been created and needs to be executed
-   */
-  readonly onDidRequestExecutePlan: vscode.Event<ExecutePlanRequest>;
-
-  /**
-   * Opens the creator mode interface
-   * @returns A Promise that resolves when the interface is opened
-   */
-  openCreatorOverlay(): Promise<void>;
-
-  /**
-   * Closes the creator mode interface
-   * @returns A Promise that resolves when the interface is closed
-   */
-  closeCreatorOverlay(): Promise<void>;
-
-  changeState(state: CreatorModeState): Promise<void>;
-
-  /**
-   * Creates a new task in creator mode
-   * @param task The task details
-   * @returns A Promise that resolves when the task is created
-   */
-  createTask(task: CreatorTaskRequest): Promise<void>;
-
-  /**
-   * Disposes of resources used by the creator mode
-   */
-  dispose(): void;
-}
-
-/**
- * Represents a request to create a new task in creator mode
- */
-export interface CreatorTaskRequest {
-  /**
-   * The text of the task
-   */
-  initialPrompt: string;
-
-  /**
-   * The plan that the AI generated with the user
-   */
-  plan: string;
-
-  /**
-   * The file path of the new task?
-   * TODO: are we going to pass this through?
-   */
-  // filePath?: string;
-
-  /**
-   * Optional base64-encoded images to include with the task
-   * TODO: are we doing images?
-   */
-  images?: string[];
-}
-
-/**
- * Represents a request to execute a plan
- */
-export interface ExecutePlanRequest {
-  /**
-   * The path to the file containing the plan
-   */
-  // filePath?: string;
-
-  /**
-   * Optional code to include in the plan execution
-   */
-  // code?: string;
-
-  /**
-   * Additional context for the plan execution
-   */
-  plan?: string;
-}
 
 /**
  * The format for all of the messages that come from the webview
@@ -122,12 +25,10 @@ export interface WebMessageOutgoing {
 /**
  * Implementation of the Creator Mode API
  * Manages the creator mode interface state and events
- * TODO: GO OVER ALL OF THESE FUNCTIONS - THEY ARE PLACEHOLDERS AND THEY NEED MODIFYING
  */
 export class PearAICreatorMode implements IPearAICreatorMode {
   // Private event emitters
   private readonly _onDidChangeCreatorModeState = new vscode.EventEmitter<CreatorModeState>();
-  private readonly _onDidRequestNewTask = new vscode.EventEmitter<CreatorTaskRequest>();
   private readonly _onDidRequestExecutePlan = new vscode.EventEmitter<ExecutePlanRequest>();
 
   // The abort token we can send to the LLM
@@ -135,7 +36,7 @@ export class PearAICreatorMode implements IPearAICreatorMode {
 
   // Public events
   public readonly onDidChangeCreatorModeState = this._onDidChangeCreatorModeState.event;
-  public readonly onDidRequestNewTask = this._onDidRequestNewTask.event;
+
   public readonly onDidRequestExecutePlan = this._onDidRequestExecutePlan.event;
 
 
@@ -153,7 +54,6 @@ export class PearAICreatorMode implements IPearAICreatorMode {
     // Add emitters to disposables
     this._disposables.push(
       this._onDidChangeCreatorModeState,
-      this._onDidRequestNewTask,
       this._onDidRequestExecutePlan
     );
   }
@@ -171,7 +71,8 @@ export class PearAICreatorMode implements IPearAICreatorMode {
         break;
       case "OVERLAY_CLOSED": 
         await vscode.commands.executeCommand("workbench.action.closeCreatorView");
-
+      case "OVERLAY_CLOSED_CREATOR_ACTIVE":
+        await vscode.commands.executeCommand("workbench.action.progressCreatorToNextStage");
 
     }
 
@@ -186,7 +87,7 @@ export class PearAICreatorMode implements IPearAICreatorMode {
     try {
       // Execute the command to open the creator mode interface
       await vscode.commands.executeCommand('workbench.action.toggleCreatorView');
-      this._onDidChangeCreatorModeState.fire(true);
+      // this._onDidChangeCreatorModeState.fire(true);
     } catch (error) {
       console.error('Failed to open creator mode:', error);
       throw error;
@@ -201,25 +102,12 @@ export class PearAICreatorMode implements IPearAICreatorMode {
     try {
       // Close the creator mode interface
       await vscode.commands.executeCommand("workbench.action.closeCreatorView");
-      this._onDidChangeCreatorModeState.fire(false);
+      // this._onDidChangeCreatorModeState.fire(false);
     } catch (error) {
       console.error('Failed to close creator mode:', error);
       throw error;
     }
   }
-
-  /**
-   * Whenever we have a new task to execute, this create task method should be called
-   * It will fire an event, which the roo code extension will listen to, then it will execute the task
-   */
-  public async createTask(task: CreatorTaskRequest): Promise<void> {
-
-
-    this._onDidRequestNewTask.fire(task);
-    // TODO: Go trigger creator mode view 
-    // TODO: run animation to close overlay
-  }
-
 
   public async handleIncomingWebViewMessage(msg: WebViewMessageIncoming, send: (messageType: string, payload: Record<string, unknown>) => string): Promise<void> {
     assert(!!msg.messageId || !!msg.messageType, "Message ID or type missing :(");
@@ -229,9 +117,6 @@ export class PearAICreatorMode implements IPearAICreatorMode {
         console.dir('GOT ProcessLLM');
 
         const { messages } = msg.payload;
-
-        const abortController = new AbortController();
-        this.cancelToken = abortController.signal;
 
         const gen = this._messenger.invoke(
           "llm/streamChat",
@@ -244,14 +129,11 @@ export class PearAICreatorMode implements IPearAICreatorMode {
             }
           }
         );
+
         let completeResponse = "";
         let next = await gen.next();
 
         while (!next.done) {
-          // if (!activeRef.current) {
-          //   abortController.abort();
-          //   break;
-          // }
           completeResponse += next.value.content;
           send("planCreationStream", {
             plan: completeResponse,
@@ -274,7 +156,7 @@ export class PearAICreatorMode implements IPearAICreatorMode {
     } else if (msg.messageType === "SubmitPlan") {
       console.dir(`MSG PAYLOAD TEXT FOR SUBMITPLAN: ${msg.payload.text}`);
       this._onDidRequestExecutePlan.fire(msg.payload); // sends off the request to the roo code extension to execute the plan
-      vscode.commands.executeCommand("workbench.action.enterCreatorMode")// sets the "view" in vscode to the creator view
+      this.changeState("OVERLAY_CLOSED_CREATOR_ACTIVE");
 
       // TODO: handle being inside of the "creator mode" whilst still having access to all of the shizz
     } else if (msg.messageType === "Close") {
